requirements.txt
flask==3.0.0
anthropic==0.39.0
gunicorn==21.2.0
python-dotenv==1.0.0
apscheduler==3.10.4
feedparser==6.0.11
praw==7.7.1
resend==2.0.0
requests==2.31.0
beautifulsoup4==4.12.3
